+++ climate_categories/__init__.py
@@ -36,6 +36,7 @@ def _read_pickle_nh(name) -> Categorization:
 # do this explicitly to help static analysis tools
 IPCC1996 = _read_pickle_hier("IPCC1996")
 IPCC2006 = _read_pickle_hier("IPCC2006")
+IPCC2006_PRIMAP = _read_pickle_hier("IPCC2006_PRIMAP")
 CRFDI = _read_pickle_hier("CRFDI")
 CRFDI_class = _read_pickle_hier("CRFDI_class")
 BURDI = _read_pickle_hier("BURDI")