+++ climate_categories/tests/test_search.py
@@ -5,6 +5,7 @@ def test_search():
     assert climate_categories.find_code("1A") == {
         climate_categories.IPCC1996["1.A"],
         climate_categories.IPCC2006["1.A"],
+        climate_categories.IPCC2006_PRIMAP["1.A"],
         climate_categories.CRFDI["1.A"],
         climate_categories.CRFDI_class["1.A"],
         climate_categories.BURDI["1.A"],